+++ linux-loader/src/lib.rs
@@ -36,7 +36,6 @@ pub fn run(args: Vec<String>, envs: Vec<String>, rootfs: Arc<dyn FileSystem>) ->
     let inode = rootfs.root_inode().lookup(&args[0]).unwrap();
     let data = inode.read_as_vec().unwrap();
     let (entry, sp) = loader.load(&proc.vmar(), &data, args, envs).unwrap();
-
     thread
         .start(entry, sp, 0, 0, spawn)
         .expect("failed to start main thread");
@@ -89,68 +88,6 @@ fn spawn(thread: Arc<Thread>) {
 }
 
 #[cfg(target_arch = "mips")]
-use mips::registers::cp0;
-
-#[cfg(target_arch = "mips")]
-use mips::addr::*;
-
-#[cfg(target_arch = "mips")]
-use mips::paging::PageTable as MIPSPageTable;
-
-#[cfg(target_arch = "mips")]
-fn get_root_page_table_ptr() -> usize {
-    extern "C" {
-        fn _root_page_table_ptr();
-    }
-    unsafe { *(_root_page_table_ptr as *mut usize) }
-}
-
-#[cfg(target_arch = "mips")]
-pub fn handle_user_page_fault(thread: &Arc<Thread>, addr: usize) -> bool {
-    let virt_addr = VirtAddr::new(addr);
-    let root_table = unsafe { &mut *(get_root_page_table_ptr() as *mut MIPSPageTable) };
-    let tlb_result = root_table.lookup(addr);
-    let flags = MMUFlags::WRITE;
-    match tlb_result {
-        Ok(tlb_entry) => {
-            trace!(
-                "PhysAddr = {:x}/{:x}",
-                tlb_entry.entry_lo0.get_pfn() << 12,
-                tlb_entry.entry_lo1.get_pfn() << 12
-            );
-
-            let tlb_valid = if virt_addr.page_number() & 1 == 0 {
-                tlb_entry.entry_lo0.valid()
-            } else {
-                tlb_entry.entry_lo1.valid()
-            };
-            if !tlb_valid {
-                // if !thread.vm.lock().handle_page_fault(addr) {
-                //     return false;
-                // }
-                match thread.proc().vmar().handle_page_fault(addr, flags) {
-                    Ok(()) => {}
-                    Err(_e) => {
-                        return false;
-                    }
-                }
-            }
-
-            tlb_entry.write_random();
-            true
-        }
-        Err(()) => match thread.proc().vmar().handle_page_fault(addr, flags) {
-            Ok(()) => {
-                return true;
-            }
-            Err(_e) => {
-                return false;
-            }
-        },
-    }
-}
-
-#[cfg(target_arch = "mips")]
 fn spawn(thread: Arc<Thread>) {
     let vmtoken = thread.proc().vmar().table_phys();
     let future = async move {
@@ -162,16 +99,16 @@ fn spawn(thread: Arc<Thread>) {
             let mut exit = false;
             let trap_num = cx.cause;
             match trap_num {
-                _ if InterruptManager::is_page_fault(trap_num) => {
-                    let addr = cp0::bad_vaddr::read_u32() as usize;
-                    if !handle_user_page_fault(&thread, addr) {
-                        // TODO: SIGSEGV
-                        panic!("page fault handle failed");
-                    }
-                }
-                _ if InterruptManager::is_syscall(trap_num) => {
-                    exit = handle_syscall(&thread, &mut cx).await
-                }
+                // _ if InterruptManager::is_page_fault(trap_num) => {
+                //     let addr = cp0::bad_vaddr::read_u32() as usize;
+                //     if !handle_user_page_fault(&thread, addr) {
+                //         // TODO: SIGSEGV
+                //         panic!("page fault handle failed");
+                //     }
+                // }
+                // _ if InterruptManager::is_syscall(trap_num) => {
+                //     exit = handle_syscall(&thread, &mut cx).await
+                // }
                 _ => panic!("not supported interrupt from user mode. {:#x?}", cx),
             }
             thread.end_running(cx);
